This commit adds new buttons to the controls, play, stop, pause, step-forwad, step-backwards. The buttons have been color coded and icons for each have been added to SVG library, components/CustonIcons.ts.
With this commit the gui now has the required controls options for simple interactive clean controls. The
buttons have been labeled and grouped according to use. The backend hooks are not current available and
the main goal was to determine the best layout and create the required components.

 On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	*deleted:    components/CustomIcon/svg_icon.py
	*modified:   components/custom_icon/CustomIcon.ts
	*modified:   interactive-clean-gui.py